Extract `useOrganizationListInView` to separate file
